break out info bubble into its own component
