+++ packages/desktop-client/src/components/settings/Upcoming.tsx
@@ -8,9 +8,9 @@ import { type SyncedPrefs } from 'loot-core/types/prefs';
 
 import { useFeatureFlag } from '../../hooks/useFeatureFlag';
 import { useSyncedPref } from '../../hooks/useSyncedPref';
-import { SvgInformationCircle } from '../../icons/v2';
 import { type CSSProperties, theme } from '../../style';
 import { Button } from '../common/Button2';
+import { InfoBubble } from '../common/InfoBubble';
 import { Select } from '../common/Select';
 import { Text } from '../common/Text';
 import { View } from '../common/View';
@@ -62,7 +62,10 @@ export function UpcomingLengthSettings() {
                   style={selectButtonStyle}
                 />
               </View>
-              <InfoCircle />
+              <InfoBubble
+                label="Only the first instance of a recurring transaction
+          will be shown."
+              />
             </View>
           </Column>
         </View>
@@ -93,50 +96,6 @@ export function UpcomingLengthSettings() {
           {options.find(x => x.value === upcomingLength)?.label ?? '1 Week'})
         </Trans>
       </Button>
-    </View>
-  );
-}
-
-function InfoCircle() {
-  const location = useLocation();
-  const [visible, setVisible] = useState(location.hash === '#info');
-
-  return visible ? (
-    <View style={{ userSelect: 'none' }}>
-      <SvgInformationCircle
-        style={{ height: '15px', cursor: 'pointer' }}
-        onClick={() => setVisible(false)}
-        onMouseLeave={() => setVisible(false)}
-      />
-      <View
-        style={{
-          position: 'absolute',
-          left: '20px',
-          top: '-20px',
-          color: 'white',
-          backgroundColor: 'black',
-          borderRadius: '5px',
-          borderStyle: 'solid',
-          borderWidth: '1px',
-          borderColor: 'white',
-          padding: '5px',
-          width: '200px',
-          zIndex: 300,
-        }}
-      >
-        <Text>
-          <strong>Only</strong> the first instance of a recurring transaction
-          will be shown.
-        </Text>
-      </View>
-    </View>
-  ) : (
-    <View style={{ userSelect: 'none' }}>
-      <SvgInformationCircle
-        style={{ height: '15px', cursor: 'pointer' }}
-        onClick={() => setVisible(true)}
-        onMouseOver={() => setTimeout(() => setVisible(true), 500)}
-      />
     </View>
   );
 }
